requirements.txt
absl-py==2.1.0
altair==6.0.0
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
awscli==1.18.69
bcrypt==3.1.7
blinker==1.8.2
botocore==1.16.19
cachetools==7.0.5
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.5
click==8.1.7
cloud-init==24.4.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.0
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.14
distlib==0.4.0
distro==1.9.0
distro-info==0.23+ubuntu1.1
docutils==0.16
entrypoints==0.3
fabric==2.5.0
fastapi==0.115.11
filelock==3.25.2
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
git-filter-repo==2.45.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.14.0
httpx==0.27.2
hyperlink==19.0.0
idna==2.8
imageio==2.35.1
importlib-metadata==1.5.0
incremental==16.10.1
invoke==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.9.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==3.4.1
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
macaroonbakery==1.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==4.2.0
mysql-connector-python==9.3.0
namex==0.0.8
narwhals==2.18.0
netifaces==0.10.4
numpy==1.26.4
oauthlib==3.1.0
olefile==0.46
openai==1.51.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.3.3
paramiko==2.6.0
pexpect==4.6.0
pillow==10.4.0
platformdirs==4.9.4
protobuf==4.25.4
psutil==5.5.1
pyarrow==23.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pycurl==7.43.0.2
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyotp==2.9.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-discovery==1.1.3
python-dotenv==1.2.2
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
requests==2.32.5
requests-unixsocket==0.2.0
rich==13.7.1
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
scipy==1.14.1
seaborn==0.13.2
SecretStorage==2.3.1
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
sos==4.8.2
SQLAlchemy==2.0.44
ssh-import-id==5.10
starlette==0.46.0
stevedore==5.7.0
streamlit==1.55.0
systemd-python==234
tenacity==9.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tf-keras-vis==0.8.7
toml==0.10.2
tornado==6.5.5
tqdm==4.66.5
Twisted==18.9.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.6.3
uvicorn==0.34.2
virtualenv==21.2.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wadllib==1.3.3
watchdog==6.0.0
Werkzeug==3.0.3
wheel==0.34.2
wrapt==1.16.0
zipp==1.0.0
zope.interface==4.7.1
